+++ anchor/programs/games/src/instructions/bounty/mod.rs
@@ -1,3 +1,7 @@
 pub mod publish;
+pub mod fund;
+pub mod renew;
 
 pub use publish::*;
+pub use fund::*;
+pub use renew::*;