+++ anchor/programs/games/src/lib.rs
@@ -9,7 +9,7 @@ pub mod state;
 
 use instructions::*;
 
-declare_id!("74arRDDazQJzSQRhm7VonhyhRnNrwBGZE4dyhNva5z8p");
+declare_id!("LuckhEzDRjC8wrPcXQyiK8Vdj5nVuurtfNtq6PQsirw");
 
 #[program]
 pub mod games {
@@ -52,21 +52,35 @@ pub mod games {
     }
 
     // ------------------------ GAME_MODE ------------------------
-    pub fn add_game_mode(ctx: Context<InitializeGameMode>, _mode_seed: String, settings: GameMode) -> Result<()> {
+    pub fn add_game_mode(ctx: Context<InitializeGameMode>, _mode_seed: String, settings: GameModeSettings) -> Result<()> {
         ctx.accounts.mode.game = ctx.accounts.game.key();
         game_mode::upsert::verify_and_set(&mut ctx.accounts.mode, settings)
     }
 
-    pub fn update_game_mode(ctx: Context<UpdateGameMode>, _mode_seed: String, settings: GameMode) -> Result<()> {
+    pub fn update_game_mode(ctx: Context<UpdateGameMode>, _mode_seed: String, settings: GameModeSettings) -> Result<()> {
         game_mode::upsert::verify_and_set(&mut ctx.accounts.mode, settings)
     }
 
     pub fn close_game_mode(_ctx: Context<CloseGameMode>, _mode_seed: String) -> Result<()> { Ok(()) }
 
     // ------------------------ BOUNTY ------------------------
-    pub fn issue_bounty(ctx: Context<InitializeBounty>, settings: Bounty) -> Result<()> {
+    pub fn issue_bounty(ctx: Context<InitializeBounty>, settings: BountySettings) -> Result<()> {
+        ctx.accounts.bounty.owner = ctx.accounts.supplier.key();
         ctx.accounts.bounty.gem = ctx.accounts.gem.key();
         ctx.accounts.bounty.task = ctx.accounts.task.key();
+        ctx.accounts.bounty.trader = ctx.accounts.trader.key();
+
         bounty::publish::new_bounty(&mut ctx.accounts.bounty, settings)
+    }
+
+    pub fn fund_bounty(ctx: Context<VaultLoad>, amount: u64) -> Result<()> {
+        let available = bounty::fund::vault_load(&ctx, amount)?;
+        bounty::fund::gems_issued(&mut ctx.accounts.bounty, available)?;
+
+        Ok(())
+    }
+
+    pub fn renew_bounty(ctx: Context<RenewBounty>, settings: BountySettings) -> Result<()> {
+        bounty::renew::existent_bounty(&mut ctx.accounts.bounty, settings)
     }
 }