+++ anchor/src/games-exports/helpers.ts
@@ -1,3 +1,4 @@
+import { BN } from '@coral-xyz/anchor';
 import { GAME_NAME_LEN } from './constants';
 
 export function encodeName(name: string) {
@@ -10,4 +11,29 @@ export function encodeName(name: string) {
 
 export function decodeName(name: number[]) {
   return new TextDecoder().decode(Uint8Array.from(name));
+}
+
+export function toBigInt(value: number, decimals: number) {
+  if (isNaN(value)) return BigInt(0);
+  if (isNaN(decimals) || decimals === 0) return BigInt(value);
+  if (decimals < 0) throw new Error('Decimals must be a positive number');
+
+  if (value < 1) return BigInt(value * 10 ** decimals);
+  return BigInt(value) * BigInt(10 ** decimals);
+}
+
+export function toBN(value: number, decimals: number) {
+  return new BN(toBigInt(value, decimals).toString());
+}
+
+export function fromBigInt(value: bigint, decimals: number) {
+  if (!value) return 0;
+  if (isNaN(decimals) || decimals === 0) return Number(value);
+  if (decimals < 0) throw new Error('Decimals must be a positive number');
+
+  return Number(value) / 10 ** decimals;
+}
+
+export function fromBN(value: BN, decimals: number) {
+  return fromBigInt(BigInt(value.toString()), decimals);
 }