+++ anchor/src/games-exports/pda/bounty.ts
@@ -1,13 +1,26 @@
 import { type Cluster, PublicKey } from '@solana/web3.js';
-import { BOUNTY_SEED, getGamesProgramId } from '..';
+import { getGamesProgramId, BOUNTY_SEED, VAULT_SEED } from '..';
 
 export function getBountyPDA(
-  gem: PublicKey,
   task: PublicKey,
+  gem: PublicKey,
+  trader: PublicKey,
   cluster?: Cluster
 ) {
   return PublicKey.findProgramAddressSync(
-    [Buffer.from(BOUNTY_SEED, 'utf8'), gem.toBytes(), task.toBytes()],
+    [
+      Buffer.from(BOUNTY_SEED, 'utf8'),
+      task.toBytes(),
+      gem.toBytes(),
+      trader.toBytes(),
+    ],
+    getGamesProgramId(cluster)
+  )[0];
+}
+
+export function getBountyVaultPDA(bounty: PublicKey, cluster?: Cluster) {
+  return PublicKey.findProgramAddressSync(
+    [Buffer.from(VAULT_SEED, 'utf8'), bounty.toBytes()],
     getGamesProgramId(cluster)
   )[0];
 }