+++ scripts/src/main.ts
@@ -3,6 +3,7 @@ config();
 
 import { CLUSTERS, createConnection } from './utils';
 import { CreateAndStockpileGem } from './tokens';
+import { CreateAndStoreTrades } from './tokens/luckyshot/create';
 
 const { CLUSTER } = process.env;
 if (!CLUSTER) throw new Error('CLUSTER is required');
@@ -17,9 +18,18 @@ const connection = createConnection(cluster);
 
 CreateAndStockpileGem(connection, cluster)
   .then(() => {
-    console.log('Token created and stockpiled');
+    console.log('Gem created and stockpiled');
   })
   .catch((err) => {
     console.log('Error creating and stockpiling gem');
+    console.error(err);
+  });
+
+CreateAndStoreTrades(connection, cluster)
+  .then(() => {
+    console.log('Trader created and store filled');
+  })
+  .catch((err) => {
+    console.log('Error creating and storing trades');
     console.error(err);
   });